Resonaate v1.4.0

+++ tests/conftest.py
@@ -1,33 +1,39 @@
-# pylint: disable=invalid-name, attribute-defined-outside-init, no-self-use
+# pylint: disable=invalid-name, attribute-defined-outside-init
 # Standard Library Imports
 import logging
 import os
 import sys
+from datetime import datetime
 from unittest.mock import create_autospec
+
 # Third Party Imports
-from numpy import asarray
 import pytest
+from numpy import asarray
+
 # Resonaate Imports
 try:
-    from resonaate.data.resonaate_database import ResonaateDatabase
-    from resonaate.data.observation import Observation
+    # RESONAATE Imports
     from resonaate.agents.estimate_agent import EstimateAgent
     from resonaate.agents.sensing_agent import SensingAgent
     from resonaate.common.behavioral_config import BehavioralConfig
-    from resonaate.filters.unscented_kalman_filter import UnscentedKalmanFilter
+    from resonaate.data.importer_database import ImporterDatabase
+    from resonaate.data.observation import Observation
+    from resonaate.data.resonaate_database import ResonaateDatabase
+    from resonaate.dynamics.special_perturbations import SpecialPerturbations
+    from resonaate.estimation.sequential.unscented_kalman_filter import UnscentedKalmanFilter
     from resonaate.parallel import getRedisConnection, isMaster, resetMaster
     from resonaate.parallel.job import Job
+    from resonaate.physics.time.stardate import ScenarioTime, datetimeToJulianDate
     from resonaate.scenario.clock import ScenarioClock
-    from resonaate.physics.time.stardate import ScenarioTime
+    from resonaate.scenario.config.geopotential_config import GeopotentialConfig
+    from resonaate.scenario.config.perturbations_config import PerturbationsConfig
     from resonaate.sensors.sensor_base import Sensor
     from resonaate.tasking.decisions.decision_base import Decision
     from resonaate.tasking.engine.centralized_engine import CentralizedTaskingEngine
     from resonaate.tasking.metrics.metric_base import Metric
     from resonaate.tasking.rewards.reward_base import Reward
 except ImportError as error:
-    raise Exception(
-        f"Please ensure you have appropriate packages installed:\n {error}"
-    ) from error
+    raise Exception(f"Please ensure you have appropriate packages installed:\n {error}") from error
 
 
 FIXTURE_DATA_DIR = os.path.join(
@@ -44,14 +50,13 @@ class BaseTestCase:
 
     importer_db_path = "db/importer.sqlite3"
     shared_db_path = "db/shared.sqlite3"
-    maneuver_db_path = "db/maneuver.sqlite3"
     json_init_path = "json/config/init_messages"
     json_rso_truth = "json/rso_truth"
     json_sensor_truth = "json/sat_sensor_truth"
 
 
 @pytest.fixture(autouse=True)
-def patchMissingEnvVariables(monkeypatch):
+def _patchMissingEnvVariables(monkeypatch):
     """Automatically delete each environment variable, if set.
 
     Args:
@@ -75,25 +80,25 @@ def getTestLoggerObject():
     logger = logging.getLogger("Unit Test Logger")
     logger.setLevel(logging.DEBUG)
     handler = logging.StreamHandler(stream=sys.stdout)
-    handler.setFormatter(
-        logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-    )
+    handler.setFormatter(logging.Formatter("%(asctime)s - %(name)s - %(levelname)s - %(message)s"))
     logger.addHandler(handler)
     return logger
 
 
-@pytest.fixture(scope="function", name="redis")
+@pytest.fixture(name="redis")
 def getRedisInstance():
     """Setup and destroy an instance of Redis key-value store."""
     redis_conn = getRedisConnection()
     isMaster()
     yield redis_conn
-    redis_conn.flushall()
     resetMaster()
+    if redis_conn:
+        redis_conn.flushall()
+        redis_conn.close()
 
 
-@pytest.fixture(name="reset_db")
-def resetDatabase():
+@pytest.fixture(name="reset_shared_db")
+def _resetDatabase():
     """Reset the database tables to avoid data integrity errors.
 
     Note:
@@ -102,6 +107,31 @@ def resetDatabase():
     """
     yield
     ResonaateDatabase.getSharedInterface().resetData(tables=ResonaateDatabase.VALID_DATA_TYPES)
+
+
+@pytest.fixture(name="reset_importer_db")
+def _resetImporterDatabase():
+    """Reset the database tables to avoid data integrity errors.
+
+    Note:
+        This fixture should be utilized any time a :class:`.ScenarioClock` object is instanitated so that the "epochs"
+        table is reset.
+    """
+    yield
+    ImporterDatabase.getSharedInterface().resetData(tables=ImporterDatabase.VALID_DATA_TYPES)
+
+
+@pytest.fixture(name="database")
+def getDataInterface():
+    """Create common, non-shared DB object for all tests.
+
+    Yields:
+        :class:`.ResonaateDatabase`: properly constructed DB object
+    """
+    # Create & yield instance.
+    shared_interface = ResonaateDatabase.getSharedInterface()
+    yield shared_interface
+    shared_interface.resetData(ResonaateDatabase.VALID_DATA_TYPES)
 
 
 @pytest.fixture(scope="class", name="mocked_clock")
@@ -120,12 +150,7 @@ def getMockedEstimateObject():
     estimate = create_autospec(EstimateAgent)
     estimate.nominal_filter.pred_p = asarray([[4, 23], [1, 67]])
     estimate.nominal_filter.est_p = asarray([[1, 1], [2, 4]])
-    estimate.nominal_filter.cross_cvr = asarray(
-        [
-            [3, 1, 12],
-            [12, 344, 2]
-        ]
-    )
+    estimate.nominal_filter.cross_cvr = asarray([[3, 1, 12], [12, 344, 2]])
     estimate.initial_covariance = asarray([[1, 1], [3, 2]])
     estimate.last_observed_at = 2458454.0
     return estimate
@@ -135,13 +160,7 @@ def getMockedEstimateObject():
 def getMockedSensorObject():
     """Create a mocked :class:`.Sensor` object."""
     sensor = create_autospec(Sensor)
-    sensor.r_matrix = asarray(
-        [
-            [7.0e-4, 0.0, 0.0],
-            [0.0, 6.5e-4, 0.0],
-            [0.0, 0.0, 8.0e-4]
-        ]
-    )
+    sensor.r_matrix = asarray([[7.0e-4, 0.0, 0.0], [0.0, 6.5e-4, 0.0], [0.0, 0.0, 8.0e-4]])
     sensor.delta_boresight = 4.0
     sensor.slew_rate = 2.0
     return sensor
@@ -151,13 +170,7 @@ def getMockedSensorObject():
 def getMockedSensingAgentObject(mocked_sensor):
     """Create a mocked :class:`.SensingAgent` object."""
     sensing_agent = create_autospec(SensingAgent)
-    mocked_sensor.r_matrix = asarray(
-        [
-            [7.0e-4, 0.0, 0.0],
-            [0.0, 6.5e-4, 0.0],
-            [0.0, 0.0, 8.0e-4]
-        ]
-    )
+    mocked_sensor.r_matrix = asarray([[7.0e-4, 0.0, 0.0], [0.0, 6.5e-4, 0.0], [0.0, 0.0, 8.0e-4]])
     mocked_sensor.delta_boresight = 4.0
     mocked_sensor.slew_rate = 2.0
 
@@ -192,7 +205,7 @@ def getMockedValidJobObject():
     job.retval = {
         "reward_matrix": [0, 0, 1],
         "visibility": [1, 0, 1],
-        "observations": [create_autospec(Observation)]
+        "observations": [create_autospec(Observation)],
     }
 
     return job
@@ -230,19 +243,50 @@ def getMockedFilterObject():
     return mocked_filter
 
 
+GEOPOTENTIAL_CONFIG = {"model": "egm96.txt", "degree": 4, "order": 4}
+
+PERTURBATIONS_CONFIG = {"third_bodies": ["sun", "moon"]}
+
+TEST_START_JD = datetimeToJulianDate(datetime(2018, 12, 1, 12))
+
+
+@pytest.fixture(name="geopotential_config")
+def getGeopotentialConfig():
+    """Return a :class:`.GeopotentialConfig` object based on :attr:`.GEOPOTENTIAL_CONFIG`."""
+    config = GeopotentialConfig()
+    config.readConfig(GEOPOTENTIAL_CONFIG)
+    return config
+
+
+@pytest.fixture(name="perturbations_config")
+def getPerturbationsConfig():
+    """Return a :class:`.PerturbationsConfig` object based on :attr:`.PERTURBATIONS_CONFIG`."""
+    config = PerturbationsConfig()
+    config.readConfig(PERTURBATIONS_CONFIG)
+    return config
+
+
+@pytest.fixture(name="dynamics")
+def getDynamics(perturbations_config, geopotential_config):
+    """Return a :class:`.SpecialPerturbations` object based on configurations."""
+    return SpecialPerturbations(TEST_START_JD, geopotential_config, perturbations_config)
+
+
 def pytest_addoption(parser):
     """Add command line options."""
-    parser.addoption(
-        "--runslow", action="store_true", default=False, help="run slow tests"
-    )
+    parser.addoption("--runslow", action="store_true", default=False, help="run slow tests")
 
 
 def pytest_configure(config):
-    """Configure pytest options without an *.ini file."""
+    """Configure pytest options without an .ini file."""
     config.addinivalue_line("markers", "slow: mark test as slow to run")
-    config.addinivalue_line("markers", "integration: mark test as an integration test")
+    config.addinivalue_line("markers", "scenario: mark test as a scenario integration test")
+    config.addinivalue_line("markers", "event: mark test as an event integration test")
+    config.addinivalue_line("markers", "service: mark test as a service layer integration test")
     config.addinivalue_line("markers", "realtime: mark test as using real time propagation")
-    config.addinivalue_line("markers", "importer: mark test as using real time propagation")
+    config.addinivalue_line(
+        "markers", "importer: mark test as using imported data rather than propgation"
+    )
     config.addinivalue_line("markers", "datafiles: creates tmpdirs for required data")
